Write results to file
